+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 0.0.4
+
+- Fix the lone leaf handling.
+- Panic section in IntoIterator impl in case of the wrong hash length.
+
 # Version 0.0.3
 
 - Add sha2 and sha3 benchmarks with 100,000 leaves.